requirements.txt
psycopg2-binary
redis
django
django-allauth
dj-database-url
djangorestframework
django-filter
django-redis
whitenoise
gunicorn
base
axios